...v1

+++ src/main/java/org/example/iec61850datatypes/common/Timestamp.java
@@ -5,6 +5,6 @@ import org.example.iec61850datatypes.Data;
 
 @lombok.Data
 public class Timestamp extends Data {
-    private double value;
+    private int value;
 
 }
